+++ apps/front/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## main
 
-- state storage with hash
+- fix real elevation when querying campToCamp
 - fix mapFlyer with other country fitBound
 - fix realIndex for DirectionMap
 - fix contextual menu debounce
-- search engine ui/ux into AutoComplete component
-- add favicon
+- fix pegman with map bearing != 0
+- feature add SearchFeature extra properties
+- feature state storage with hash
+- feature pegman update position on click
+- enhancement search engine ui/ux into AutoComplete component
+- enhancement add favicon
 
 ## 0.0.1